Renamed T3 set of tests to TransactionInferenceTest.
